JAVA-2754 move mapper classes to a dedicated package
