+++ quickstart/src/main/java/com/datastax/oss/quarkus/FruitDaoReactive.java
@@ -18,7 +18,7 @@ package com.datastax.oss.quarkus;
 import com.datastax.oss.driver.api.mapper.annotations.Dao;
 import com.datastax.oss.driver.api.mapper.annotations.Select;
 import com.datastax.oss.driver.api.mapper.annotations.Update;
-import com.datastax.oss.quarkus.runtime.api.reactive.MutinyMappedReactiveResultSet;
+import com.datastax.oss.quarkus.runtime.api.reactive.mapper.MutinyMappedReactiveResultSet;
 import io.smallrye.mutiny.Uni;
 
 @Dao
