+++ runtime/src/main/java/com/datastax/oss/quarkus/runtime/api/reactive/mapper/MutinyMappedResultProducer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.datastax.oss.quarkus.runtime.api.reactive;
+package com.datastax.oss.quarkus.runtime.api.reactive.mapper;
 
 import com.datastax.dse.driver.api.core.cql.reactive.ReactiveResultSet;
 import com.datastax.dse.driver.internal.mapper.reactive.DefaultMappedReactiveResultSet;
